app/requirements.txt
absl-py==2.1.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.100.0
fastapi-jwt-auth==0.5.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
idna==3.10
Jinja2==3.1.4
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.2
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.1
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
protobuf==4.25.5
psutil==6.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pydantic==2.0.2
pydantic_core==2.1.2
Pygments==2.18.0
PyJWT==1.7.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rich==13.9.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.16.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.27.0
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
termcolor==2.5.0
tf_keras==2.17.0
thop==0.1.1.post2209072238
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.0.42
ultralytics-thop==2.0.9
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
